% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_genes}
\alias{collapse_genes}
\title{Select probe sets to represent genes}
\usage{
collapse_genes(x, id_col = NULL, genes = NULL, method = c("maxMean",
  "minMean"), drop = character(), df = FALSE)
}
\arguments{
\item{x}{A matrix, data frame or data table.}

\item{id_col}{Column name of identifiers column (key column).}

\item{genes}{A vector of length \code{nrow(x)} of the "genes" identifiers or
a mapping of probes to genes (an object coercible to a data table with
"probes" identifiers in the first column and "genes" identifiers in the
second).  In this last case, \code{x} should have its "probes" identifiers in
the \code{id_col} column or in its row names.}

\item{method}{Method to choose the "probe set" representative of a "gene":
the one with maximum row mean ('maxMean') or with minimum row mean
('minMean'). Defaults to \code{"maxMean"}.}

\item{drop}{Columns to drop (before summarizing and from result).}

\item{df}{Return a data frame instead of a data table.  Defaults to
\code{FALSE}.}
}
\value{
Table with unique observations for each "gene".
}
\description{
Select a "probe set" to represent each "gene" choosing those with highest or
lowest mean expression level (see details in \code{method}).
}
\seealso{
\code{\link{collapse_replicates}}
}

