% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_replicates}
\alias{collapse_replicates}
\title{Collapse technical replicates to unique values}
\usage{
collapse_replicates(x, id_col = NULL, replicates = NULL,
  method = c("mean", "median"), drop = character(), df = FALSE)
}
\arguments{
\item{x}{A matrix, data frame or data table.}

\item{id_col}{Column name of identifiers column (key column).}

\item{replicates}{A vector of length \code{nrow(x)} of replicates
identifiers.  Optional, if \code{id_col} isn't in \code{x} or to substitute
\code{id_col}.}

\item{method}{Method to summarize rows.  Defaults to \code{"median"}.}

\item{drop}{Columns to drop (before summarizing and from result).}

\item{df}{Return a data frame instead of a data table.  Defaults to
\code{FALSE}.}
}
\value{
Table with unique identifiers and single values for each variable and
observation.
}
\description{
Summarize replicated data to unique values per identifier using the the
median or the mean values.
}
\note{
Uses median without checking or data.tables' internaly optimized mean.
}
\seealso{
\code{\link{collapse_genes}}
}

