% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{fmedian}
\alias{fmedian}
\title{Median of a numeric vector}
\usage{
fmedian(x)
}
\arguments{
\item{x}{}
}
\value{
The median.
}
\description{
Calculates the median without checking the input or removing NAs.
}
\seealso{
\code{\link[stats]{median}}
}

