% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Transform dataset to a standard table format}
\usage{
format_table(x, key_col = key(x), key_data = NULL, sort = TRUE,
  df = FALSE)
}
\arguments{
\item{x}{A matrix, data frame or data table.}

\item{key_col}{Character vector of one or more column names which is passed
to \code{\link[data.table]{setkey}}.  If \code{x} has a key, it can be
ommited.}

\item{key_data}{A vector of length \code{nrow(x)} to be used as key.  If the
(first) column name in \code{key_col} is already a column of \code{x}, it
will be substituted by \code{key_data} and the table reordered.  Otherwise,
and if \code{key_data} is  missing, \code{row.names(x)} will be used instead.}

\item{sort}{Wheter to sort by \code{key_col}.  Defaults to \code{TRUE}.}

\item{df}{Return a data frame instead of a data table.  Defaults to
\code{FALSE}.}
}
\value{
A data table (or data frame) based on \code{x} and ordered by \code{key_col}.
}
\description{
Transforms a matrix, data frame or data table to an ordered data table with
the key as the first column.  The key column can be passed, selected by its
column name or extracted from row names.
}
\examples{
data(geneData, geneCovariate, package='Biobase')
gData <- head(geneData, n=5)

gData
class(gData)

gData <- format_table(gData, 'Probeset')
gData
class(gData)
key(gData)

gData <- format_table(gData, key_data=row.names(geneData)[6:10], sort=FALSE)
gData
key(gData)

}

